refactor(configurator): split drawing page modules
